Move check-cfg diagnostic logic into it's own module
